REfactor RosterSnapshotCharacter to a shared directory.
