+++ src/fsd/1-pages/input-roster-snapshots/roster-snapshots-unit-diff-detailed.tsx
@@ -8,6 +8,7 @@ import { RarityMapper } from '@/fsd/5-shared/model/mappers/rarity.mapper';
 import { abilityIcons } from '@/fsd/5-shared/ui/ability-icons';
 import { MiscIcon, RarityIcon, StarsIcon, RankIcon } from '@/fsd/5-shared/ui/icons';
 import { tacticusIcons } from '@/fsd/5-shared/ui/icons/icon-list';
+import { ISnapshotCharacter, ISnapshotMachineOfWar } from '@/fsd/5-shared/ui/unit-portrait';
 
 import { CharactersService } from '@/fsd/4-entities/character';
 import { EquipmentService, IEquipment } from '@/fsd/4-entities/equipment';
@@ -15,7 +16,7 @@ import { MowsService } from '@/fsd/4-entities/mow';
 
 import { RosterSnapshotShowVariableSettings } from '@/fsd/3-features/view-settings/model';
 
-import { ISnapshotCharacter, ISnapshotMachineOfWar, ISnapshotUnitDiff } from './models';
+import { ISnapshotUnitDiff } from './models';
 
 interface ProgressionRowProps {
     diffFlag: boolean;
