+++ src/fsd/1-pages/input-roster-snapshots/roster-snapshots-unit.tsx
@@ -1,11 +1,10 @@
+/* eslint-disable import-x/no-internal-modules */
+import { UnitPortrait, ISnapshotCharacter, ISnapshotMachineOfWar } from '@/fsd/5-shared/ui/unit-portrait';
+
 import { CharactersService } from '@/fsd/4-entities/character';
 import { MowsService } from '@/fsd/4-entities/mow';
 
-// eslint-disable-next-line import-x/no-internal-modules
 import { RosterSnapshotShowVariableSettings } from '@/fsd/3-features/view-settings/model';
-
-import { ISnapshotCharacter, ISnapshotMachineOfWar } from './models';
-import { RosterSnapshotCharacter } from './roster-snapshot-character';
 
 interface Props {
     showShards: RosterSnapshotShowVariableSettings;
@@ -37,7 +36,7 @@ export const RosterSnapshotsUnit: React.FC<Props> = ({
         <div className="flex h-[200px] w-[110px] overflow-hidden rounded-lg border border-gray-200 bg-white shadow-xl transition duration-300 hover:border-blue-500/50 hover:shadow-2xl dark:border-gray-700/50 dark:bg-gray-800">
             <div className="relative flex h-full w-full flex-shrink-0 items-center justify-center bg-gray-100 p-1 dark:bg-gray-900">
                 {char !== undefined && staticChar !== undefined && (
-                    <RosterSnapshotCharacter
+                    <UnitPortrait
                         showShards={showShards}
                         showMythicShards={showMythicShards}
                         showXpLevel={showXpLevel}
@@ -50,7 +49,7 @@ export const RosterSnapshotsUnit: React.FC<Props> = ({
                     />
                 )}
                 {mow !== undefined && staticMow !== undefined && (
-                    <RosterSnapshotCharacter
+                    <UnitPortrait
                         showShards={showShards}
                         showMythicShards={showMythicShards}
                         showXpLevel={showXpLevel}
