+++ src/fsd/5-shared/ui/unit-portrait/index.ts
@@ -1,0 +1,4 @@
+export { UnitPortrait } from './unit-portrait';
+export { UnitPortraitAssetsProvider } from './unit-portrait-assets-provider';
+export { useUnitPortraitAssets } from './unit-portrait.hooks';
+export type { ISnapshotCharacter, ISnapshotMachineOfWar, UnitPortraitAssets } from './unit-portrait.model';
