+++ src/fsd/5-shared/ui/unit-portrait/unit-portrait-assets-context.tsx
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+import { UnitPortraitAssets } from './unit-portrait.model';
+
+export const UnitPortraitAssetContext = createContext<UnitPortraitAssets | undefined>(undefined);
